Move showPublishNote and publishNoteActions to note controller
