+++ lib/web/note/actions.js
@@ -8,7 +8,6 @@ const markdownpdf = require('markdown-pdf')
 const moment = require('moment')
 const querystring = require('querystring')
 
-
 exports.getInfo = function getInfo (req, res, note) {
   const body = note.content
   const extracted = models.Note.extractMeta(body)
