+++ lib/web/note/router.js
@@ -1,9 +1,6 @@
 'use strict'
 
 const Router = require('express').Router
-
-const response = require('../../response')
-
 const { markdownParser } = require('../utils')
 
 const router = module.exports = Router()
@@ -18,9 +15,9 @@ router.post('/new', markdownParser, noteController.createFromPOST)
 // post new note with content and alias
 router.post('/new/:noteId', markdownParser, noteController.createFromPOST)
 // get publish note
-router.get('/s/:shortid', response.showPublishNote)
+router.get('/s/:shortid', noteController.showPublishNote)
 // publish note actions
-router.get('/s/:shortid/:action', response.publishNoteActions)
+router.get('/s/:shortid/:action', noteController.publishNoteActions)
 // get publish slide
 router.get('/p/:shortid', slide.showPublishSlide)
 // publish slide actions
